Merge `release/2.3.0` branch into `master` branch

+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/utils/UniqueIdentifierGenerator.kt
@@ -39,7 +39,7 @@ object UniqueIdentifierGenerator {
             // was a long and use that as an identifier.
             uniqueIdentifier as? Long
         } else {
-            val newUniqueIdentifier = secureRandom.nextLong()
+            val newUniqueIdentifier = secureRandom.nextInt().toLong()
             parent.setTag(key, newUniqueIdentifier)
             newUniqueIdentifier
         }
